chat_service/pyproject.toml
[tool.poetry]
name = "chat-service"
version = "0.1.0"
description = ""
authors = ["Victor Momodu <victoramomodu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.115.0"}
python-multipart = "^0.0.11"
openai = "^1.50.2"
qdrant-client = "^1.11.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
